requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pdf2image==1.14.0
pillow==10.2.0
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
PyPDF2==3.0.1
pytesseract==0.3.7
SQLAlchemy==2.0.28
typing_extensions==4.10.0
Werkzeug==3.0.1
